+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "vscode-tilt" extension will be documented in this fi
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [Unreleased]
+
+### Added
+
+- Tilt's log stream to extension's Output channel
+
 ## [0.0.3] - 2025-09-14
 
 ### Fixed